Move `ScrollPositionContext` into `LocationSlide` folder
